requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
werkzeug==2.3.7
requests==2.31.0
watchdog>=6.0.0
